Adding explicit variable conversion in ocl_medianFilter() to suppress build errors on Linux & Mac

+++ modules/imgproc/src/smooth.cpp
@@ -2023,8 +2023,8 @@ static bool ocl_medianFilter(InputArray _src, OutputArray _dst, int m)
 
     Size imgSize = _src.size();
     bool useOptimized = (1 == cn) &&
-                        imgSize.width >= localsize[0] * 8  &&
-                        imgSize.height >= localsize[1] * 8 &&
+                        (size_t)imgSize.width >= localsize[0] * 8  &&
+                        (size_t)imgSize.height >= localsize[1] * 8 &&
                         (ocl::Device::getDefault().isIntel());
 
     cv::String kname = format( useOptimized ? "medianFilter%d_u" : "medianFilter%d", m) ;
